requirements.txt
numpy==1.21.5
openai==0.27.7
openai_whisper==20230314
opencv_python==4.7.0.72
Pillow==9.5.0
pytesseract==0.3.10
scipy==1.8.0
streamlit==1.22.0
tqdm==4.65.0
imageio==2.30.0
imageio-ffmpeg==0.4.8
